Feature/add workplace create base class for enter workplace address (#1117)
* feature | add workplace create base class for enter workplace address

* fix routing links

* use for loop to generate form controls html resulting in slimmer markup

* create WIP enter workplace address base class

* update base class

* extend enterworkplace address component and use on registration flow
