+++ src/app/features/workplace-find-and-select/enter-workplace-address/enter-workplace-address.component.spec.ts
@@ -1,0 +1,23 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { EnterWorkplaceAddress } from './enter-workplace-address';
+
+describe('EnterWorkplaceAddressComponent', () => {
+  let component: EnterWorkplaceAddress;
+  let fixture: ComponentFixture<EnterWorkplaceAddress>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [EnterWorkplaceAddress],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(EnterWorkplaceAddress);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
